+++ filament/src/PostProcessManager.cpp
@@ -201,6 +201,7 @@ struct MaterialInfo {
 
 static const MaterialInfo sMaterialList[] = {
         { "bilateralBlur",              MATERIAL(BILATERALBLUR) },
+        { "bilateralBlurBentNormals",   MATERIAL(BILATERALBLURBENTNORMALS) },
         { "blitHigh",                   MATERIAL(BLITHIGH) },
         { "blitLow",                    MATERIAL(BLITLOW) },
         { "blitMedium",                 MATERIAL(BLITMEDIUM) },
@@ -221,6 +222,7 @@ static const MaterialInfo sMaterialList[] = {
         { "fxaa",                       MATERIAL(FXAA) },
         { "mipmapDepth",                MATERIAL(MIPMAPDEPTH) },
         { "sao",                        MATERIAL(SAO) },
+        { "saoBentNormals",             MATERIAL(SAOBENTNORMALS) },
         { "separableGaussianBlur",      MATERIAL(SEPARABLEGAUSSIANBLUR) },
         { "taa",                        MATERIAL(TAA) },
         { "vsmMipmap",                  MATERIAL(VSMMIPMAP) },
@@ -399,7 +401,8 @@ FrameGraphId<FrameGraphTexture> PostProcessManager::screenSpaceAmbientOcclusion(
     // (see en.wikipedia.org/wiki/Gaussian_filter)
     // More intuitively, 2q is the width of the filter in pixels.
     BilateralPassConfig config = {
-            .bilateralThreshold = options.bilateralThreshold
+            .bentNormals = options.bentNormals,
+            .bilateralThreshold = options.bilateralThreshold,
     };
 
     float sampleCount{};
@@ -452,7 +455,11 @@ FrameGraphId<FrameGraphTexture> PostProcessManager::screenSpaceAmbientOcclusion(
     struct SSAOPassData {
         FrameGraphId<FrameGraphTexture> depth;
         FrameGraphId<FrameGraphTexture> ssao;
+        FrameGraphId<FrameGraphTexture> ao;
+        FrameGraphId<FrameGraphTexture> bn;
     };
+
+    const bool computeBentNormals = options.bentNormals;
 
     const bool highQualityUpsampling =
             options.upsampling >= View::QualityLevel::HIGH && options.resolution < 1.0f;
@@ -467,10 +474,21 @@ FrameGraphId<FrameGraphTexture> PostProcessManager::screenSpaceAmbientOcclusion(
                 data.ssao = builder.createTexture("SSAO Buffer", {
                         .width = desc.width,
                         .height = desc.height,
-                        .depth = 1,
+                        .depth = computeBentNormals ? 2u : 1u,
                         .type = Texture::Sampler::SAMPLER_2D_ARRAY,
-                        .format = (lowPassFilterEnabled || highQualityUpsampling) ? TextureFormat::RGB8 : TextureFormat::R8
+                        .format = (lowPassFilterEnabled || highQualityUpsampling || computeBentNormals) ?
+                                TextureFormat::RGB8 : TextureFormat::R8
                 });
+
+                if (computeBentNormals) {
+                    data.ao = builder.createSubresource(data.ssao, "SSAO attachment", { .layer = 0 });
+                    data.bn = builder.createSubresource(data.ssao, "Bent Normals attachment", { .layer = 1 });
+                    data.ao = builder.write(data.ao, FrameGraphTexture::Usage::COLOR_ATTACHMENT);
+                    data.bn = builder.write(data.bn, FrameGraphTexture::Usage::COLOR_ATTACHMENT);
+                } else {
+                    data.ao = data.ssao;
+                    data.ao = builder.write(data.ao, FrameGraphTexture::Usage::COLOR_ATTACHMENT);
+                }
 
                 // Here we use the depth test to skip pixels at infinity (i.e. the skybox)
                 // Note that we have to clear the SAO buffer because blended objects will end-up
@@ -478,18 +496,17 @@ FrameGraphId<FrameGraphTexture> PostProcessManager::screenSpaceAmbientOcclusion(
                 // The bilateral filter in the blur pass will ignore pixels at infinity.
 
                 data.depth = builder.read(data.depth, FrameGraphTexture::Usage::DEPTH_ATTACHMENT);
-                data.ssao = builder.write(data.ssao, FrameGraphTexture::Usage::COLOR_ATTACHMENT);
                 builder.declareRenderPass("SSAO Target", {
-                        .attachments = { .color = { data.ssao }, .depth = data.depth },
+                        .attachments = { .color = { data.ao, data.bn }, .depth = data.depth },
                         .clearColor = { 1.0f },
-                        .clearFlags = TargetBufferFlags::COLOR
+                        .clearFlags = TargetBufferFlags::COLOR0 | TargetBufferFlags::COLOR1
                 });
             },
             [=](FrameGraphResources const& resources,
                     auto const& data, DriverApi& driver) {
                 auto depth = resources.getTexture(data.depth);
                 auto ssao = resources.getRenderPassInfo();
-                auto const& desc = resources.getDescriptor(data.ssao);
+                auto const& desc = resources.getDescriptor(data.depth);
 
                 // estimate of the size in pixel of a 1m tall/wide object viewed from 1m away (i.e. at z=-1)
                 const float projectionScale = std::min(
@@ -512,7 +529,10 @@ FrameGraphId<FrameGraphTexture> PostProcessManager::screenSpaceAmbientOcclusion(
                         0.0, 0.0, 0.0, 1.0
                 }};
 
-                auto& material = getPostProcessMaterial("sao");
+                auto& material = computeBentNormals ?
+                            getPostProcessMaterial("saoBentNormals") :
+                            getPostProcessMaterial("sao");
+
                 FMaterialInstance* const mi = material.getMaterialInstance();
                 mi->setParameter("depth", depth, {
                         .filterMin = SamplerMinFilter::NEAREST_MIPMAP_NEAREST });
@@ -584,7 +604,7 @@ FrameGraphId<FrameGraphTexture> PostProcessManager::screenSpaceAmbientOcclusion(
                 config);
 
         ssao = bilateralBlurPass(fg, ssao, { 0, config.scale }, cameraInfo.zf,
-                highQualityUpsampling ? TextureFormat::RGB8 : TextureFormat::R8,
+                (highQualityUpsampling || computeBentNormals) ? TextureFormat::RGB8 : TextureFormat::R8,
                 config);
     }
 
@@ -601,6 +621,8 @@ FrameGraphId<FrameGraphTexture> PostProcessManager::bilateralBlurPass(
     struct BlurPassData {
         FrameGraphId<FrameGraphTexture> input;
         FrameGraphId<FrameGraphTexture> blurred;
+        FrameGraphId<FrameGraphTexture> ao;
+        FrameGraphId<FrameGraphTexture> bn;
     };
 
     auto& blurPass = fg.addPass<BlurPassData>("Separable Blur Pass",
@@ -621,13 +643,24 @@ FrameGraphId<FrameGraphTexture> PostProcessManager::bilateralBlurPass(
                 assert_invariant(depth);
                 depth = builder.read(depth, FrameGraphTexture::Usage::DEPTH_ATTACHMENT);
 
+                if (config.bentNormals) {
+                    data.ao = builder.createSubresource(data.blurred, "SSAO attachment", { .layer = 0 });
+                    data.bn = builder.createSubresource(data.blurred, "Bent Normals attachment", { .layer = 1 });
+                    data.ao = builder.write(data.ao, FrameGraphTexture::Usage::COLOR_ATTACHMENT);
+                    data.bn = builder.write(data.bn, FrameGraphTexture::Usage::COLOR_ATTACHMENT);
+                } else {
+                    data.ao = data.blurred;
+                    data.ao = builder.write(data.ao, FrameGraphTexture::Usage::COLOR_ATTACHMENT);
+                }
+
                 // Here we use the depth test to skip pixels at infinity (i.e. the skybox)
                 // We need to clear the buffers because we are skipping pixels at infinity (skybox)
                 data.blurred = builder.write(data.blurred, FrameGraphTexture::Usage::COLOR_ATTACHMENT);
+
                 builder.declareRenderPass("Blurred target", {
-                        .attachments = { .color = { data.blurred }, .depth = depth },
+                        .attachments = { .color = { data.ao, data.bn }, .depth = depth },
                         .clearColor = { 1.0f },
-                        .clearFlags = TargetBufferFlags::COLOR
+                        .clearFlags = TargetBufferFlags::COLOR0 | TargetBufferFlags::COLOR1
                 });
             },
             [=](FrameGraphResources const& resources,
@@ -653,7 +686,9 @@ FrameGraphId<FrameGraphTexture> PostProcessManager::bilateralBlurPass(
                 uint32_t kGaussianCount = gaussianKernel(kGaussianSamples,
                         config.kernelSize, config.standardDeviation);
 
-                auto& material = getPostProcessMaterial("bilateralBlur");
+                auto& material = config.bentNormals ?
+                        getPostProcessMaterial("bilateralBlurBentNormals") :
+                        getPostProcessMaterial("bilateralBlur");
                 FMaterialInstance* const mi = material.getMaterialInstance();
                 mi->setParameter("ssao", ssao, { /* only reads level 0 */ });
                 mi->setParameter("axis", axis / float2{desc.width, desc.height});